fix exports and naming
