+++ src/services/swap/index.ts
@@ -4,8 +4,8 @@ import { SarosBaseService, SarosConfig } from '../base';
 import { FeeCalculator } from './fees';
 import { VolatilityManager } from './volatility';
 import { BinArrayRange } from './bin-range';
-import { SwapExecutor } from './execution';
-import { MAX_BIN_CROSSINGS, SCALE_OFFSET } from '../../constants';
+import * as spl from '@solana/spl-token';
+import { BIN_ARRAY_SIZE, MAX_BIN_CROSSINGS, SCALE_OFFSET, WRAP_SOL_PUBKEY } from '../../constants';
 import {
   DLMMPairAccount,
   QuoteParams,
@@ -25,19 +25,23 @@ import {
 import { PoolServiceError } from '../pools/errors';
 import { SwapServiceError } from './errors';
 import { BinArrayManager } from '../pools/bin-manager';
+import { addSolTransferInstructions, addCloseAccountInstruction } from '../../utils/transaction';
+import { getTokenProgram, getUserVaultInfo } from '../../utils/vaults';
 
 export class SwapService extends SarosBaseService {
   private volatilityManager: VolatilityManager;
-  private swapExecutor: SwapExecutor;
 
   constructor(config: SarosConfig) {
     super(config);
     this.volatilityManager = new VolatilityManager();
-    this.swapExecutor = new SwapExecutor(this.lbProgram, this.hooksProgram, this.connection);
   }
 
   private async calculateInOutAmount(params: QuoteParams) {
-    const { amount, swapForY, pair, isExactInput } = params;
+    const {
+      amount,
+      pair,
+      options: { swapForY, isExactInput },
+    } = params;
     try {
       //@ts-ignore
       const pairInfo: DLMMPairAccount = await this.lbProgram.account.pair.fetch(pair);
@@ -385,28 +389,184 @@ export class SwapService extends SarosBaseService {
 
   // only swap and getQuote are public
   public async swap(params: SwapParams): Promise<Transaction> {
-    return this.swapExecutor.executeSwap(params);
+    const {
+      tokenMintX,
+      tokenMintY,
+      amount,
+      otherAmountOffset,
+      options: { swapForY, isExactInput },
+
+      pair,
+      hook,
+      payer,
+    } = params;
+
+    //@ts-ignore
+    const pairInfo: DLMMPairAccount = await this.lbProgram.account.pair.fetch(pair);
+    if (!pairInfo) throw PoolServiceError.PoolNotFound;
+
+    const currentBinArrayIndex = BinArrayManager.calculateBinArrayIndex(pairInfo.activeId);
+
+    const surroundingIndexes = [
+      currentBinArrayIndex - 1,
+      currentBinArrayIndex,
+      currentBinArrayIndex + 1,
+    ];
+
+    const binArrayAddresses = await Promise.all(
+      surroundingIndexes.map(async (idx) =>
+        BinArrayManager.getBinArrayAddress(idx, pair, this.lbProgram.programId)
+      )
+    );
+
+    const binArrayAccountsInfo = await this.connection.getMultipleAccountsInfo(binArrayAddresses);
+
+    const validIndexes = surroundingIndexes.filter((_, i) => binArrayAccountsInfo[i]);
+
+    if (validIndexes.length < 2) {
+      throw new Error('No valid bin arrays found for the pair');
+    }
+
+    let binArrayLowerIndex: number;
+    let binArrayUpperIndex: number;
+    if (validIndexes.length === 2) {
+      [binArrayLowerIndex, binArrayUpperIndex] = validIndexes;
+    } else {
+      const activeOffset = pairInfo.activeId % BIN_ARRAY_SIZE;
+      const [first, second, third] = validIndexes;
+      [binArrayLowerIndex, binArrayUpperIndex] =
+        activeOffset < BIN_ARRAY_SIZE / 2 ? [first, second] : [second, third];
+    }
+
+    const binArrayLower = BinArrayManager.getBinArrayAddress(
+      binArrayLowerIndex,
+      pair,
+      this.lbProgram.programId
+    );
+
+    const binArrayUpper = BinArrayManager.getBinArrayAddress(
+      binArrayUpperIndex,
+      pair,
+      this.lbProgram.programId
+    );
+
+    const [tokenProgramX, tokenProgramY] = await Promise.all([
+      getTokenProgram(tokenMintX, this.connection),
+      getTokenProgram(tokenMintY, this.connection),
+    ]);
+
+    const latestBlockHash = await this.connection.getLatestBlockhash();
+    const tx = new Transaction({
+      feePayer: payer,
+      blockhash: latestBlockHash.blockhash,
+      lastValidBlockHeight: latestBlockHash.lastValidBlockHeight,
+    });
+
+    const associatedPairVaultX = spl.getAssociatedTokenAddressSync(
+      tokenMintX,
+      pair,
+      true,
+      tokenProgramX
+    );
+
+    const associatedPairVaultY = spl.getAssociatedTokenAddressSync(
+      tokenMintY,
+      pair,
+      true,
+      tokenProgramY
+    );
+
+    // Use the centralized getUserVaultInfo function to create user vault accounts if needed
+    const associatedUserVaultX = await getUserVaultInfo(
+      { tokenMint: tokenMintX, payer, transaction: tx },
+      this.connection
+    );
+
+    const associatedUserVaultY = await getUserVaultInfo(
+      { tokenMint: tokenMintY, payer, transaction: tx },
+      this.connection
+    );
+
+    // Handle wrapped SOL transfers
+    if (tokenMintY.equals(WRAP_SOL_PUBKEY) || tokenMintX.equals(WRAP_SOL_PUBKEY)) {
+      const isNativeY = tokenMintY.equals(WRAP_SOL_PUBKEY);
+      const associatedUserVault = isNativeY ? associatedUserVaultY : associatedUserVaultX;
+
+      if ((isNativeY && !swapForY) || (!isNativeY && swapForY)) {
+        addSolTransferInstructions(tx, payer, associatedUserVault, amount);
+      }
+    }
+
+    // Add swap instruction
+    const swapInstructions = await this.lbProgram.methods
+      .swap(
+        new BN(amount.toString()),
+        new BN(otherAmountOffset.toString()),
+        swapForY,
+        isExactInput ? { exactInput: {} } : { exactOutput: {} }
+      )
+      .accountsPartial({
+        pair: pair,
+        binArrayLower: binArrayLower,
+        binArrayUpper: binArrayUpper,
+        tokenVaultX: associatedPairVaultX,
+        tokenVaultY: associatedPairVaultY,
+        userVaultX: associatedUserVaultX,
+        userVaultY: associatedUserVaultY,
+        tokenMintX: tokenMintX,
+        tokenMintY: tokenMintY,
+        tokenProgramX,
+        tokenProgramY,
+        user: payer,
+        hook: hook || null,
+        hooksProgram: this.hooksProgram.programId,
+      })
+      .remainingAccounts([
+        { pubkey: pair, isWritable: false, isSigner: false },
+        { pubkey: binArrayLower, isWritable: false, isSigner: false },
+        { pubkey: binArrayUpper, isWritable: false, isSigner: false },
+      ])
+      .instruction();
+
+    tx.add(swapInstructions);
+
+    // Handle wrapped SOL account closing
+    if (tokenMintY.equals(WRAP_SOL_PUBKEY) || tokenMintX.equals(WRAP_SOL_PUBKEY)) {
+      const isNativeY = tokenMintY.equals(WRAP_SOL_PUBKEY);
+      const associatedUserVault = isNativeY ? associatedUserVaultY : associatedUserVaultX;
+      if ((isNativeY && swapForY) || (!isNativeY && !swapForY)) {
+        addCloseAccountInstruction(tx, associatedUserVault, payer);
+      }
+    }
+
+    return tx;
   }
 
   public async getQuote(params: QuoteParams, poolMetadata: PoolMetadata): Promise<QuoteResponse> {
     try {
       const { baseToken, quoteToken } = poolMetadata;
+      const {
+        slippage,
+        pair,
+        options: { swapForY, isExactInput },
+      } = params;
+
       const data = await this.calculateInOutAmount(params);
       const { amountIn, amountOut } = data;
 
       let maxAmountIn = amountIn;
       let minAmountOut = amountOut;
 
-      if (params.isExactInput) {
-        minAmountOut = getMinOutputWithSlippage(amountOut, params.slippage);
+      if (isExactInput) {
+        minAmountOut = getMinOutputWithSlippage(amountOut, slippage);
       } else {
-        maxAmountIn = getMaxInputWithSlippage(amountIn, params.slippage);
+        maxAmountIn = getMaxInputWithSlippage(amountIn, slippage);
       }
 
       const { maxAmountOut } = await this.getMaxAmountOutWithFee(
-        params.pair,
+        pair,
         amountIn,
-        params.swapForY,
+        swapForY,
         baseToken.decimals,
         quoteToken.decimals
       );
@@ -416,8 +576,8 @@ export class SwapService extends SarosBaseService {
       return {
         amountIn: amountIn,
         amountOut: amountOut,
-        amount: params.isExactInput ? maxAmountIn : minAmountOut,
-        otherAmountOffset: params.isExactInput ? minAmountOut : maxAmountIn,
+        amount: isExactInput ? maxAmountIn : minAmountOut,
+        otherAmountOffset: isExactInput ? minAmountOut : maxAmountIn,
         priceImpact: priceImpact,
       };
     } catch (error) {
