+++ src/services/swap/tests/swap.test.ts
@@ -17,8 +17,7 @@ describe('getQuote with bigint support', () => {
     const quote = await lbServices.getQuote({
       amount: largeAmount,
       pair: new PublicKey('9P3N4QxjMumpTNNdvaNNskXu2t7VHMMXtePQB72kkSAk'),
-      swapForY: true,
-      isExactInput: true,
+      options: { swapForY: true, isExactInput: true },
       slippage: 10,
     });
