+++ src/types/swap.ts
@@ -1,15 +1,16 @@
 import { PublicKey } from '@solana/web3.js';
 
+interface Options {
+  swapForY: boolean; // true = X to Y, false = Y to X
+  isExactInput: boolean;
+}
+
 export interface SwapParams {
-  // TODO: consider rename to inputToken/outputToken - tokenMintX/Y is unclear which is which
   tokenMintX: PublicKey;
   tokenMintY: PublicKey;
   amount: bigint;
-  /** Slippage protection threshold - unclear naming, consider minAmountOut/maxAmountIn */
-  otherAmountOffset: bigint;
-  /** Direction: true = X to Y, false = Y to X */
-  swapForY: boolean;
-  isExactInput: boolean;
+  options: Options;
+  otherAmountOffset: bigint; // slippage protection threshold
   pair: PublicKey;
   hook: PublicKey;
   payer: PublicKey;
@@ -18,18 +19,14 @@ export interface SwapParams {
 export interface QuoteParams {
   pair: PublicKey;
   amount: bigint;
-  /** Direction: true = X to Y, false = Y to X */
-  swapForY: boolean;
-  isExactInput: boolean;
-  /** Slippage percentage (0-100) */
-  slippage: number;
+  options: Options;
+  slippage: number; // allowed slippage as a percentage (0-100)
 }
 
 export interface QuoteResponse {
   amountIn: bigint;
   amountOut: bigint;
-  /** Price impact as percentage */
-  priceImpact: number;
+  priceImpact: number; // price impact as percentage (0-100)
   amount: bigint;
   /** Slippage-adjusted threshold - unclear naming, consider minAmountOut/maxAmountIn */
   otherAmountOffset: bigint;
